Make kWh interpolation optional
It is currently rather unreliable. The energy values are given
with a 0.01 kWh accuracy, so if we interpolate it using the
power value, the next energy value refresh (command class 50) or
other update will cancel out the interpolation, making the used
energy value drop.

This could be solved by relying more on the interpolated energy
value and less on the one indicated in register 50. That requires
a lot of extra administration and makes errors possible.

Rather, we are now just using ZWAVE_REFRESH_INTERVAL = 1000 so
that the energy value is refreshed every second. The power value
is ignored because ZWAVE_INTERPOLATE = False.